adding deleteProduct in admin dashboard

+++ server/services/productServices.js
@@ -187,13 +187,16 @@ const deleteProduct = async (productId) => {
 
   if (!product) return "Product not found... ";
 
-  // if (product.image.public_id) {
-  //   const destroyResponse = await cloudinary.uploader.destroy(
-  //     product.image.public_id
-  //   );
-  const deletedProduct = await Product.findByIdAndDelete(productId);
+  if (product.image.public_id) {
+    const destroyResponse = await cloudinary.uploader.destroy(
+      product.image.public_id
+    );
+    if (destroyResponse) {
+      const deletedProduct = await Product.findByIdAndDelete(productId);
 
-  return deletedProduct;
+      return deletedProduct;
+    }
+  }
 };
 
 // get by id
